adding some things like players spawn
