--- Assets/EnemiesManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fca2093e63602184a8168303cd103a0c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
