--- Assets/NewLayer 1.terrainlayer.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a928fd54ff70e294b87dfd4477d3f746
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 8574412962073106934
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
