--- Assets/NewLayer 3.terrainlayer.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bcfc5bda055dc0b458c66fcc5f7dd790
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 8574412962073106934
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
