--- Assets/NewLayer.terrainlayer.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 314569470888c7a4382a9a63efcc8107
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 8574412962073106934
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
