+++ Assets/Resources/PhotonPrefabs/DonteOnline.prefab
@@ -89,6 +89,7 @@ GameObject:
   - component: {fileID: 7886396514305899175}
   - component: {fileID: -954626089738484577}
   - component: {fileID: -4892904066769281840}
+  - component: {fileID: 3554999803534188748}
   m_Layer: 12
   m_Name: DonteOnline
   m_TagString: Player
@@ -599,6 +600,24 @@ MonoBehaviour:
   m_SynchronizeRotation: 1
   m_SynchronizeScale: 0
   m_UseLocal: 1
+--- !u!114 &3554999803534188748
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3778023270062103502}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5fea29bb7c508c244a1f805a5fd3fc4d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  outlineMode: 2
+  outlineColor: {r: 1, g: 1, b: 1, a: 1}
+  outlineWidth: 2
+  precomputeOutline: 1
+  bakeKeys: []
+  bakeValues: []
 --- !u!1 &5094332642605627631
 GameObject:
   m_ObjectHideFlags: 0
