+++ Assets/_Environment/Terrains/HubWorld/New Terrain.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6f885d544289c974c804863e24c5725c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
