+++ Assets/_Environment/Terrains/HubWorld/NewLayer 2.terrainlayer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 027c05825ef32a242b9aec36a755cd36
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8574412962073106934
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
