+++ Assets/_GameController/Prefabs/PlayerSpawn.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 60203b1b469319149aec3e5d81a276de
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
