+++ Assets/_GameController/Scripts/PlayerSpawnPoint.cs
@@ -5,4 +5,8 @@ using UnityEngine;
 public class PlayerSpawnPoint : MonoBehaviour
 {
     public string spawnName;
+    void Start()
+    {
+        GetComponent<MeshRenderer>().enabled = false;
+    }
 }
