+++ Assets/_Materials/InfoRune.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9f5fa3ff04498b54182788168eccefa0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
