+++ Assets/_Protype/BolderBarrier.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 87e3d709acb284a4c8955c4c50bfecca
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
