+++ Assets/_Utils/HoverEffect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd7dc5fdeead853469305a2e3618e37a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
