+++ UserSettings/EditorUserSettings.asset
@@ -21,16 +21,16 @@ EditorUserSettings:
       value: 500301555301510d08595b2113770f4044161d7d742b22687e2a4463e0e2663a
       flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 53075053500c5909555c087a49705e4444161a787a7c7e667b784c35b6e56360
-      flags: 0
-    RecentlyUsedSceneGuid-6:
       value: 0600070554030b030f5a5b7a43270d44444f1e7f7f782562297e4a36b1b4306d
       flags: 0
+    RecentlyUsedSceneGuid-6:
+      value: 53075053500c5909555c087a49705e4444161a787a7c7e667b784c35b6e56360
+      flags: 0
     RecentlyUsedSceneGuid-7:
-      value: 07570000500d505e0e0b5e2111725a4417151c7c7c7d7f607b7d4f64b1e6646c
+      value: 5552515553015c0308570a20437b0844404f1929287872687e714865b4e3366c
       flags: 0
     RecentlyUsedSceneGuid-8:
-      value: 5552515553015c0308570a20437b0844404f1929287872687e714865b4e3366c
+      value: 07570000500d505e0e0b5e2111725a4417151c7c7c7d7f607b7d4f64b1e6646c
       flags: 0
     RecentlyUsedSceneGuid-9:
       value: 5403565054065c08595a5e75157a5b4443161a72797d71367b2d1c61b3e5623a
